Reinstated 'Kogoruhn - Extinct City of Ash and Sulfur' rule
'Kogoruhn - Extinct City of Ash and Sulfur' previously required 'Riharradroon - Hidden Path to Kogoruhn'. This dependency was removed in v2.4. 

Prior to this, the mod description advised loading Kogoruhn after Riharradroon. Demanufacturer87 confirmed in the Nexus page comments on 2023-11-06 that both mods are still compatible together. Opted to reinstate the order rule.